ref: move ScheduleEntry to the modules which need it
